# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "description": "Global catalog; available to any authenticated user.",
        "operationId": "ProductsController_list",
        "parameters": [
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List products",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "400": {
            "description": "Request body or parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the role or warehouse scope is not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "A product with this SKU already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id (UUID).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a product by id",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id (UUID).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "400": {
            "description": "Request body or parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the role or warehouse scope is not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "A product with this SKU already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a product",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "description": "Soft-deletes the product (status → ARCHIVED).",
        "operationId": "ProductsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id (UUID).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the role or warehouse scope is not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Archive a product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/barcode": {
      "get": {
        "description": "Renders the product SKU as a PNG (code128 or qr) data URI.",
        "operationId": "ProductsController_barcode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id (UUID).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "symbology",
            "required": false,
            "in": "query",
            "description": "Barcode symbology; defaults to code128.",
            "schema": {
              "example": "code128",
              "type": "string",
              "enum": [
                "code128",
                "qr"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarcodeDto"
                }
              }
            }
          },
          "400": {
            "description": "Request body or parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Generate a barcode for a product",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Inventory & Warehouse System API",
    "description": "REST API for the Inventory & Warehouse System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code.",
            "example": 403
          },
          "error": {
            "type": "string",
            "description": "Short HTTP error label.",
            "example": "Forbidden"
          },
          "message": {
            "description": "Human-readable message, or an array of validation messages for 400s.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Out of scope"
          },
          "timestamp": {
            "type": "string",
            "description": "ISO-8601 timestamp when the error was produced.",
            "example": "2026-06-15T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "Request path that produced the error.",
            "example": "/api/v1/warehouses/9c1b…"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "ProductStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "ARCHIVED"
        ],
        "description": "Product status."
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Product id (UUID).",
            "example": "9c1b2d34-5678-90ab-cdef-1234567890ab"
          },
          "name": {
            "type": "string",
            "description": "Product name.",
            "example": "Cola 330ml Can"
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping unit (unique).",
            "example": "COLA-330"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Free-text description."
          },
          "categoryId": {
            "type": "string",
            "nullable": true,
            "description": "Category id (UUID), or null."
          },
          "unit": {
            "type": "string",
            "nullable": true,
            "description": "Unit of measure."
          },
          "costPrice": {
            "type": "string",
            "description": "Cost price as a decimal string (never a float).",
            "example": "0.45"
          },
          "sellingPrice": {
            "type": "string",
            "description": "Selling price as a decimal string.",
            "example": "1.20"
          },
          "reorderLevel": {
            "type": "number",
            "description": "Reorder level.",
            "example": 50
          },
          "status": {
            "description": "Product status.",
            "example": "ACTIVE",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductStatus"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp.",
            "example": "2026-06-19T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "sku",
          "costPrice",
          "sellingPrice",
          "reorderLevel",
          "status",
          "createdAt"
        ]
      },
      "BarcodeDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "The encoded value (a product or variant SKU/barcode).",
            "example": "TSHIRT-L"
          },
          "symbology": {
            "enum": [
              "code128",
              "qr"
            ],
            "type": "string",
            "description": "Barcode symbology used.",
            "example": "code128"
          },
          "png": {
            "type": "string",
            "description": "Rendered barcode as a PNG data URI.",
            "example": "data:image/png;base64,iVBORw0KGgo…"
          }
        },
        "required": [
          "value",
          "symbology",
          "png"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name.",
            "example": "Cola 330ml Can"
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping unit (unique).",
            "example": "COLA-330"
          },
          "description": {
            "type": "string",
            "description": "Free-text description."
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "Category id (UUID).",
            "example": "9c1b2d34-5678-90ab-cdef-1234567890ab"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measure.",
            "example": "can"
          },
          "costPrice": {
            "type": "number",
            "minimum": 0,
            "description": "Cost price (used for stock valuation).",
            "example": 0.45
          },
          "sellingPrice": {
            "type": "number",
            "minimum": 0,
            "description": "Selling price.",
            "example": 1.2
          },
          "reorderLevel": {
            "type": "number",
            "minimum": 0,
            "description": "Reorder level — low-stock alert threshold.",
            "example": 50
          }
        },
        "required": [
          "name",
          "sku"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name.",
            "example": "Cola 330ml Can"
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping unit (unique).",
            "example": "COLA-330"
          },
          "description": {
            "type": "string",
            "description": "Free-text description."
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "Category id (UUID).",
            "example": "9c1b2d34-5678-90ab-cdef-1234567890ab"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measure.",
            "example": "can"
          },
          "costPrice": {
            "type": "number",
            "minimum": 0,
            "description": "Cost price (used for stock valuation).",
            "example": 0.45
          },
          "sellingPrice": {
            "type": "number",
            "minimum": 0,
            "description": "Selling price.",
            "example": 1.2
          },
          "reorderLevel": {
            "type": "number",
            "minimum": 0,
            "description": "Reorder level — low-stock alert threshold.",
            "example": 50
          },
          "status": {
            "description": "Product status.",
            "example": "ARCHIVED",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductStatus"
              }
            ]
          }
        }
      },
...
    }
  }
}