# packages/api-client/src/generated/model/barcodeDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Inventory & Warehouse System API
 * REST API for the Inventory & Warehouse System
 * OpenAPI spec version: 1.0
 */
import type { BarcodeDtoSymbology } from './barcodeDtoSymbology';

export interface BarcodeDto {
  /** The encoded value (a product or variant SKU/barcode). */
  value: string;
  /** Barcode symbology used. */
  symbology: BarcodeDtoSymbology;
  /** Rendered barcode as a PNG data URI. */
  png: string;
}